chore: address AI quality findings
